wxt(p2): popup.js becomes popup-core - logic survives with its pins, renders await their React successors
