+++ apps/caramel-extension/dom-utils.js
@@ -453,7 +453,8 @@ export function caramelCurrencySymbol() {
 
 /* ISO code for the symbol we actually parsed off the page, for the savings
  * HISTORY. The modal renders the symbol, but the popup totals the history per
- * currency through Intl.NumberFormat (popup.js:81-98), which needs a code —
+ * currency through Intl.NumberFormat (popup-core.js formatSavingsTotal),
+ * which needs a code —
  * and banking every DOM-path win as 'USD' silently added £ and € savings into
  * the dollar bucket, overstating a non-US user's lifetime total.
  * '$' stays USD: a bare dollar sign can't distinguish USD/CAD/AUD, and USD is
