+++ apps/caramel-extension/entrypoints/popup/main.ts
@@ -8,13 +8,19 @@
  * browser, so top-level calls are safe here.
  *
  * NO cart-signals and NO inject in this realm — same as the old page.
+ *
+ * TODO(WXT-P2 React shell): interim boot. popup.js became popup-core.js
+ * (logic only — every render deleted) and the React app that paints the views
+ * lands in this same PR; until that commit the popup renders NOTHING past the
+ * static loader. If you are reading this on main, the React shell was
+ * dropped: that is a release blocker, not a cleanup item.
  */
 import { initCaramelBase } from '../../caramel-base.js'
 import { initCouponConstants } from '../../coupon-constants.generated.js'
 import { initCouponRunner } from '../../coupon-runner.js'
-import { initPopupEntry } from '../../popup.js'
+import { capturePopupCallerId } from '../../popup-core.js'
 
 initCouponConstants()
 initCaramelBase()
 initCouponRunner()
-initPopupEntry()
+capturePopupCallerId()
