+++ apps/caramel-extension/tests/popup-auth-validate.test.mjs
@@ -1,20 +1,22 @@
-import { beforeAll, beforeEach, describe, expect, it } from 'vitest'
+import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
 import { initCaramelBase } from '../caramel-base.js'
 import { initCouponConstants } from '../coupon-constants.generated.js'
 import { initCouponRunner } from '../coupon-runner.js'
-import { initPopup } from '../popup.js'
+import { resolvePopupState } from '../popup-core.js'
 
 // Pins the popup's session validation: a stored token is no longer trusted
-// forever. initPopup() fires GET /api/extension/me with the bearer IN
-// PARALLEL with the coupon fetch — a real 401 clears token+user from
-// storage.sync and re-renders the logged-out variant; a 200 keeps the
-// signed-in state (and refreshes the stored user when the profile
-// changed). Network errors are NOT a logout — offline must never sign the
-// user out.
+// forever. resolvePopupState() (P2 successor to initPopup — it returns the
+// view instead of painting it) fires GET /api/extension/me with the bearer IN
+// PARALLEL with the coupon fetch — a real 401 clears token+user from storage
+// and invokes the onSessionInvalid callback (the React app's re-resolve; the
+// logged-out REPAINT half of the old pin lives with the React shell's own
+// suite); a 200 keeps the signed-in state (and refreshes the stored user when
+// the profile changed). Network errors are NOT a logout — offline must never
+// sign the user out.
 //
-// Harness mirrors popup-settings-view.test.mjs: the popup realm's own inits in
-// index.html order, one shared chrome stub, only the messaging transport +
-// storage + fetch stubbed.
+// Harness: the popup realm's own inits in entrypoints/popup/main.tsx order,
+// one shared chrome stub, only the messaging transport + storage + fetch
+// stubbed.
 let syncData
 let meResponse
 let meInit
@@ -131,7 +133,13 @@ beforeEach(() => {
     }
 })
 
-describe('popup.js initPopup — stored-token validation via /api/extension/me', () => {
+const onSessionInvalid = vi.fn()
+
+describe('popup-core resolvePopupState — stored-token validation via /api/extension/me', () => {
+    beforeEach(() => {
+        onSessionInvalid.mockClear()
+    })
+
     // /api/extension/me is declared `auth: 'session'`, and better-auth's
     // session gate accepts a website COOKIE as readily as a bearer token. That
     // is only harmless because this probe never sends one: the popup runs on a
@@ -148,7 +156,7 @@ describe('popup.js initPopup — stored-token validation via /api/extension/me',
     it('authenticates with the bearer token ALONE and never opts into sending cookies', async () => {
         meResponse = { ok: true, status: 200, json: async () => ({}) }
 
-        await initPopup()
+        await resolvePopupState(onSessionInvalid)
         await flush()
 
         expect(meInit, 'the /me probe was made').not.toBeNull()
@@ -156,15 +164,16 @@ describe('popup.js initPopup — stored-token validation via /api/extension/me',
         expect(meInit.credentials).toBeUndefined()
     })
 
-    it('a token the backend 401s clears token+user from storage and re-renders the logged-out variant', async () => {
+    it('a token the backend 401s clears token+user from storage and fires the re-resolve', async () => {
         meResponse = { ok: false, status: 401 }
-        await initPopup()
+        await resolvePopupState(onSessionInvalid)
         await flush()
         expect(syncData.token).toBeUndefined()
         expect(syncData.user).toBeUndefined()
-        const html = document.getElementById('auth-container').innerHTML
-        expect(html).toContain('Guest')
-        expect(html).not.toContain('caramel-fan')
+        // The React app registers its re-resolve here — this callback firing
+        // IS what repaints the logged-out variant (the paint itself is the
+        // React shell suite's pin).
+        expect(onSessionInvalid).toHaveBeenCalledTimes(1)
     })
 
     it('a token the backend accepts (200) keeps the signed-in state and the stored token', async () => {
@@ -173,12 +182,12 @@ describe('popup.js initPopup — stored-token validation via /api/extension/me',
             status: 200,
             json: async () => ({ username: 'caramel-fan', image: '' }),
         }
-        await initPopup()
+        const state = await resolvePopupState(onSessionInvalid)
         await flush()
         expect(syncData.token).toBe('tok-1')
-        expect(document.getElementById('auth-container').innerHTML).toContain(
-            '@caramel-fan',
-        )
+        expect(state.view).toBe('coupons')
+        expect(state.user).toEqual({ username: 'caramel-fan', image: '' })
+        expect(onSessionInvalid).not.toHaveBeenCalled()
     })
 
     it('a 200 with a changed profile refreshes the stored user', async () => {
@@ -187,7 +196,7 @@ describe('popup.js initPopup — stored-token validation via /api/extension/me',
             status: 200,
             json: async () => ({ username: 'renamed', image: 'pic.png' }),
         }
-        await initPopup()
+        await resolvePopupState(onSessionInvalid)
         await flush()
         expect(syncData.user).toEqual({ username: 'renamed', image: 'pic.png' })
         expect(syncData.token).toBe('tok-1')
@@ -197,11 +206,11 @@ describe('popup.js initPopup — stored-token validation via /api/extension/me',
         globalThis.fetch = async () => {
             throw new TypeError('Failed to fetch')
         }
-        await initPopup()
+        const state = await resolvePopupState(onSessionInvalid)
         await flush()
         expect(syncData.token).toBe('tok-1')
-        expect(document.getElementById('auth-container').innerHTML).toContain(
-            '@caramel-fan',
-        )
+        expect(state.view).toBe('coupons')
+        expect(state.user).toEqual({ username: 'caramel-fan', image: '' })
+        expect(onSessionInvalid).not.toHaveBeenCalled()
     })
 })
