+++ apps/caramel-extension/tests/popup-caller-relay.test.mjs
@@ -2,14 +2,19 @@ import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
 import { initCaramelBase } from '../caramel-base.js'
 import { initCouponConstants } from '../coupon-constants.generated.js'
 import { initCouponRunner } from '../coupon-runner.js'
-import { afterLoginSuccess, initPopupEntry } from '../popup.js'
+import {
+    afterLoginSuccess,
+    capturePopupCallerId,
+    setAfterLoginRerender,
+} from '../popup-core.js'
 
-// WXT-migration P0 characterization pins (2026-08-12): the checkout-modal
-// caller relay had ZERO coverage, and it is exactly the seam the React/WXT
-// popup rewrite disturbs — CARAMEL_CALLER_ID is read from location.search when
-// initPopupEntry() runs (popup.js, the ESM port's successor to the old
-// module-eval-time read), and its failure mode is silent: the modal's "Sign In"
-// flow simply never resumes in the originating tab.
+// WXT-migration P0 characterization pins (2026-08-12; P2-ported to popup-core
+// 2026-08-13): the checkout-modal caller relay had ZERO coverage, and it is
+// exactly the seam the React/WXT popup rewrite disturbs — CARAMEL_CALLER_ID
+// is read from location.search when capturePopupCallerId() runs (the React
+// boot's FIRST popup-core call, before anything renders), and its failure
+// mode is silent: the modal's "Sign In" flow simply never resumes in the
+// originating tab.
 //
 // The contract under pin (popup half — background.js's half is pinned in
 // background-caller-relay.test.mjs, including the round trip):
@@ -17,53 +22,17 @@ import { afterLoginSuccess, initPopupEntry } from '../popup.js'
 //     `userLoggedInFromPopup_<tabId>` (single underscore before the id — the
 //     worker parses with split('_')[1]) and closes the window ~150ms later,
 //     WITHOUT re-rendering; a dead originating tab must not prevent the close.
-//   - opened normally (toolbar), it re-renders via initPopup() and neither
-//     messages the worker nor closes.
+//   - opened normally (toolbar), it re-renders IN PLACE: the callback the
+//     React app registers via setAfterLoginRerender() fires, and neither a
+//     worker message nor a close happens. (The vanilla popup hard-called
+//     initPopup() here; the registration seam is the P2 successor, and THIS
+//     suite is what fails if the React boot forgets to register.)
 //
-// initPopupEntry() is re-run per describe-block with the URL already in place,
-// because that call is what captures the id — a pin in itself: set the URL
-// after the capture and the relay silently degrades to the toolbar branch.
-// (`CARAMEL_CALLER_ID` defaults to null, the same value URLSearchParams.get()
-// returns for an absent parameter.)
-//
-// "initPopup ran" is observed one layer down, at initPopup's own first act:
-// it awaits getActiveTabDomainRecord(), which calls
-// caramelSendMessage({action:'getActiveTabDomainRecord'}) synchronously. Under
-// ESM that call resolves to popup.js's module-local binding, so the old
-// `globalThis.initPopup = vi.fn()` swap has no seam to replace — mocking the
-// collaborator it reaches for does, and it keeps the raw
-// currentBrowser.runtime.sendMessage recorder below free to record ONLY the
-// relay's own worker message.
-
-const POPUP_DOM =
-    '<div id="loading-container"></div>' +
-    '<button id="settingsIcon" style="display:none"></button>' +
-    '<div id="auth-container"></div>'
-
-/* Never settles: initPopup awaits it, so the probe records that the render
- * started without letting the rest of the chain run under fake timers. */
-const stubs = vi.hoisted(() => ({
-    caramelSendMessage: vi.fn(() => new Promise(() => {})),
-}))
-
-vi.mock('../caramel-base.js', async importOriginal => {
-    const actual = await importOriginal()
-    return {
-        ...actual,
-        // `currentBrowser` is a live binding that initCaramelBase() assigns;
-        // a plain spread would freeze its pre-init `undefined`.
-        get currentBrowser() {
-            return actual.currentBrowser
-        },
-        caramelSendMessage: stubs.caramelSendMessage,
-    }
-})
-
-/** How many times initPopup() has started since the last reset. */
-const initPopupRuns = () =>
-    stubs.caramelSendMessage.mock.calls.filter(
-        ([message]) => message?.action === 'getActiveTabDomainRecord',
-    ).length
+// capturePopupCallerId() is re-run per describe-block with the URL already in
+// place, because that call is what captures the id — a pin in itself: set the
+// URL after the capture and the relay silently degrades to the toolbar
+// branch. (`CARAMEL_CALLER_ID` defaults to null, the same value
+// URLSearchParams.get() returns for an absent parameter.)
 
 /* Realm stub, lifted from tests/_load.mjs (installChromeStub), which the ESM
  * port retires. Permissive Proxy: any unknown property materializes as a
@@ -111,8 +80,7 @@ function installChromeStub() {
 
 // The realm is stood up ONCE — vitest gives the FILE one module registry, and
 // caramel-base's double-load guard means a second stub would be orphaned on a
-// currentBrowser nobody reads from (the same thing the old harness did across
-// its two loads). What each describe re-runs is the capture.
+// currentBrowser nobody reads from. What each describe re-runs is the capture.
 beforeAll(() => {
     installChromeStub()
     initCouponConstants()
@@ -120,20 +88,23 @@ beforeAll(() => {
     initCouponRunner()
 })
 
+/** The React app's registered re-resolve, observed instead of performed. */
+const rerender = vi.fn()
+
 const openPopupAt = urlPath => {
     history.replaceState(null, '', urlPath)
-    document.body.innerHTML = POPUP_DOM
     window.close = vi.fn()
-    initPopupEntry()
+    capturePopupCallerId()
+    setAfterLoginRerender(rerender)
 }
 
 beforeEach(() => {
-    stubs.caramelSendMessage.mockClear()
+    rerender.mockClear()
 })
 
 describe('caller relay — popup opened by the checkout modal (?callerId=42)', () => {
     beforeAll(() => {
-        openPopupAt('/index.html?isPopup=true&callerId=42')
+        openPopupAt('/popup.html?isPopup=true&callerId=42')
     })
 
     it('notifies the worker with userLoggedInFromPopup_<callerId> and closes ~150ms later, without re-rendering', () => {
@@ -152,7 +123,7 @@ describe('caller relay — popup opened by the checkout modal (?callerId=42)', (
             // parses the id with split('_')[1], so the prefix must keep its
             // single trailing underscore and contain no other one.
             expect(sent).toEqual([{ action: 'userLoggedInFromPopup_42' }])
-            expect(initPopupRuns()).toBe(0)
+            expect(rerender).not.toHaveBeenCalled()
 
             // The close is DELAYED so the message reaches the worker first.
             vi.advanceTimersByTime(149)
@@ -176,7 +147,7 @@ describe('caller relay — popup opened by the checkout modal (?callerId=42)', (
 
             vi.advanceTimersByTime(150)
             expect(window.close).toHaveBeenCalledTimes(1)
-            expect(initPopupRuns()).toBe(0)
+            expect(rerender).not.toHaveBeenCalled()
         } finally {
             vi.useRealTimers()
         }
@@ -205,10 +176,10 @@ describe('caller relay — popup opened by the checkout modal (?callerId=42)', (
 
 describe('caller relay — plain toolbar popup (no callerId)', () => {
     beforeAll(() => {
-        openPopupAt('/index.html')
+        openPopupAt('/popup.html')
     })
 
-    it('re-renders in place: initPopup runs, no worker message, no close', () => {
+    it('re-renders in place: the registered rerender runs, no worker message, no close', () => {
         vi.useFakeTimers()
         try {
             const sent = []
@@ -220,7 +191,7 @@ describe('caller relay — plain toolbar popup (no callerId)', () => {
 
             afterLoginSuccess()
 
-            expect(initPopupRuns()).toBe(1)
+            expect(rerender).toHaveBeenCalledTimes(1)
             expect(sent).toEqual([])
             vi.advanceTimersByTime(1000)
             expect(window.close).not.toHaveBeenCalled()
