+++ apps/caramel-extension/tests/popup-logout-revoke.test.mjs
@@ -5,36 +5,32 @@ import { beforeAll, beforeEach, describe, expect, it } from 'vitest'
 import { initCaramelBase } from '../caramel-base.js'
 import { initCouponConstants } from '../coupon-constants.generated.js'
 import { initCouponRunner } from '../coupon-runner.js'
-import { initPopup } from '../popup.js'
+import { signOutAndRevoke } from '../popup-core.js'
 
-// "Log out" used to be `storage.sync.remove(['token','user'])` and nothing
-// else, in all three places it appears. The bearer it forgot locally stayed
-// valid server-side for the rest of its 7-day life, so a token captured before
-// logout still authenticated afterwards — and no endpoint existed that could
-// have revoked it. DELETE /api/extension/session is that endpoint; this pins
-// that the popup actually calls it.
-//
-// The second half matters as much as the first: logging out must still work
-// when the revoke cannot be delivered. Someone offline who presses "log out"
-// has to end up logged out on this device, so the local clear is unconditional
-// — but it must not be reordered ahead of the revoke, or the token needed to
-// authenticate the revoke would already be gone.
+// Pins the popup logout: revoking the session, not just forgetting it
+// (P2-ported 2026-08-13 — the suite drives signOutAndRevoke directly; the
+// vanilla render it used to click through died with popup.js, and every
+// React view's Log out button calls exactly this function, which the
+// source-scan pin at the bottom keeps true). Logout used to be storage-only,
+// so the bearer it forgot kept authenticating for the rest of its 7-day life.
+// The revoke goes out FIRST with the stored bearer; the local clear runs
+// whether or not the revoke succeeded — offline must still sign this device
+// out.
+
 let syncData
 let requests
 let revokeResponse
 
 const flush = async () => {
-    for (let i = 0; i < 12; i++) await new Promise(r => setTimeout(r, 0))
+    for (let i = 0; i < 5; i++) {
+        await new Promise(resolve => setTimeout(resolve, 0))
+    }
 }
 
 /* Realm stub, lifted from tests/_load.mjs (installChromeStub), which the ESM
- * port retires. Permissive Proxy: any unknown property materializes as a
- * callable no-op, so a source file touching an API this suite doesn't care
- * about cannot abort it. Two deliberate exceptions, exactly as _load.mjs had
- * them — storage.*.get/set/remove invoke their callbacks like the real API
- * (empty storage), and runtime.lastError stays UNDEFINED outside a failing
- * callback, because the proxy would otherwise auto-create a truthy callable
- * that caramelSendMessage reads as a closed port. */
+ * port retires — permissive Proxy, storage callbacks invoked, lastError
+ * undefined outside a failing callback (see popup-oauth-success.test.mjs for
+ * the rationale block). */
 function installChromeStub() {
     const cache = new WeakMap()
     const wrap = target => {
@@ -72,29 +68,11 @@ function installChromeStub() {
 }
 
 beforeAll(() => {
-    document.body.innerHTML =
-        '<div id="loading-container"></div>' +
-        '<button id="settingsIcon" style="display:none"></button>' +
-        '<div id="auth-container"></div>'
-
-    // The realm's effects, in entrypoints/popup/main.ts order — the successor
-    // to the <script> list this suite used to eval.
     installChromeStub()
     initCouponConstants()
     initCaramelBase()
     initCouponRunner()
 
-    globalThis.currentBrowser.runtime.sendMessage = (message, cb) => {
-        if (message?.action === 'getActiveTabDomainRecord') {
-            cb({ url: 'https://www.example.com/cart' })
-        } else if (message?.action === 'fetchCoupons') {
-            cb({
-                coupons: [{ code: 'SAVE10', title: 'Save', status: 'valid' }],
-            })
-        } else {
-            cb(undefined)
-        }
-    }
     // The session lives in storage.LOCAL, so that is the area whose removal
     // marks "the user is signed out on this device" — instrument it, and let
     // sync (which only ever holds a pre-migration leftover) stay empty and
@@ -125,18 +103,17 @@ beforeEach(() => {
             if (revokeResponse instanceof Error) throw revokeResponse
             return revokeResponse
         }
-        // the /me probe initPopup fires in parallel
         return { ok: true, status: 200, json: async () => ({}) }
     }
 })
 
-/** Renders the signed-in popup and presses its Log out button. */
+/** Presses "Log out" the way every React view does: signOutAndRevoke with a
+ * real button element (the busy-latch contract is pinned separately in
+ * logout-feedback.test.mjs). */
 const logOut = async () => {
-    await initPopup()
-    await flush()
-    const btn = document.getElementById('logoutBtn')
-    expect(btn, 'the signed-in popup rendered a logout button').toBeTruthy()
-    btn.click()
+    const button = document.createElement('button')
+    button.textContent = 'Log out'
+    signOutAndRevoke(() => {}, button)
     await flush()
 }
 
@@ -191,17 +168,23 @@ describe('popup logout — revoking the session, not just forgetting it', () => 
     })
 
     // The popup renders THREE different signed-in views, each with its own
-    // logout button, and the tests above can only drive one of them. All three
-    // used to clear storage directly; they now share signOutAndRevoke(). A
-    // future edit that inlines the storage clear back into one of them would
-    // silently restore the un-revokable session for that view alone, and no
-    // behavioural test here would notice — so guard it at the source.
+    // logout button, and the tests above can only drive the shared function.
+    // All three used to clear storage directly; they now share
+    // signOutAndRevoke(). A future edit that inlines the storage clear back
+    // into the logic module would silently restore the un-revokable session,
+    // and no behavioural test here would notice — so guard it at the source.
+    // (The React views' buttons calling signOutAndRevoke is pinned by their
+    // own testing-library suites.)
     it('leaves no logout path that clears the session without revoking it', () => {
         const src = readFileSync(
-            join(dirname(fileURLToPath(import.meta.url)), '..', 'popup.js'),
+            join(
+                dirname(fileURLToPath(import.meta.url)),
+                '..',
+                'popup-core.js',
+            ),
             'utf8',
         )
-        // Every token-clearing site in the popup now goes through
+        // Every token-clearing site in the popup logic goes through
         // caramelClearSession(), which owns BOTH storage areas (local for the
         // session, sync only to sweep a pre-migration leftover). A raw
         // storage.*.remove of 'token' back in here would mean some path
